acds/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-generativeai==0.8.3
python-dotenv==1.0.1
websockets==12.0
python-multipart==0.0.12
requests==2.32.3
